requirements.txt
fastapi>=0.103.2,<0.104.0
uvicorn>=0.15.0
pydantic>=2.11.4   
langgraph>=0.0.10
python-multipart>=0.0.5
langgraph-studio>=0.0.1
python-dotenv>=0.19.0 